Add files via upload
This teacher assistant uses the LLM model Qwen2.5-7B to help us to generate questions (you can choose the number of questions but this code only generates 5) and hint who helps the student to answer the questions, all from a pdf document. 
It is hardly recommended to run it in Google Colab, since this code uses the "files.upload()" function to upload the pdf file.